JavaScript ShadowRealm proposal integration
This patch gives the TC39 JavaScript ShadowRealm proposal semantics in HTML.
ShadowRealms are given an environment settings object; in general,
ShadowRealms inherit their settings from the outer environment.

When modules are used in ShadowRealms, module specifiers are interpreted with
respect to the base URL of the surrounding environment, but when a module is
imported within a ShadowRealm, it is a separate copy from what may be loaded
in the surrounding environment or other ShadowRealms.

This patch implements the plan described earlier at
<https://github.com/tc39/proposal-shadowrealm/issues/225#issuecomment-592159225>.

The ShadowRealm proposal is currently at Stage 3.